Move everything data/file related to one module
